+++ crates/config-utils/src/main.rs
@@ -1,0 +1,31 @@
+use clap::Parser;
+use config_utils::template::{self, cli_args::TemplateCommand, template};
+use snafu::{ResultExt, Snafu};
+
+use crate::cli_args::{Args, Command};
+mod cli_args;
+
+#[derive(Debug, Snafu)]
+pub enum Error {
+    #[snafu(display("Failed to template file"))]
+    TemplateFile { source: template::Error },
+}
+
+type Result<T, E = Error> = std::result::Result<T, E>;
+
+#[snafu::report]
+fn main() -> Result<()> {
+    let args = Args::parse();
+
+    match args.command {
+        Command::Template(TemplateCommand {
+            file,
+            file_type,
+            dont_escape,
+        }) => {
+            template(&file, file_type.as_ref(), !dont_escape).context(TemplateFileSnafu)?;
+        }
+    }
+
+    Ok(())
+}